moved domain classes to domain
