+++ app/core/domain.py
@@ -1,8 +1,32 @@
 from dataclasses import asdict, fields
+from datetime import datetime
 from typing import Any
 
-from langchain_core.language_models.chat_models import BaseChatModel
-from pydantic.dataclasses import Field, dataclass
+from pydantic import BaseModel, Field
+from pydantic.dataclasses import dataclass
+
+
+class SearchProvider:
+    """
+    Represents a search provider.
+    """
+    GOOGLE = "google"
+    DUCKDUCKGO = "duckduckgo"
+
+
+class SearchResult(BaseModel):
+    """
+    Represents a search result from a web search engine.
+    """
+    title: str = Field(description="The title of the search result.")
+    url: str = Field(description="The URL of the search result.")
+    description: str = Field(
+        description="The description of the search result.")
+    created_at: datetime = Field(
+        description="The date and time the job was created")
+    source: str = Field(description="The source of the search result.")
+    snippet: str | None = Field(
+        description="The snippet of the search result.")
 
 
 class ModelHost:
@@ -13,6 +37,30 @@ class ModelHost:
 
 class ModelProvider:
     ANTHROPIC = "anthropic"
+
+
+class ScrapePageResult(BaseModel):
+    url: str = Field(description="The URL to scrape")
+    success: bool = Field(description="Whether the scraping was successful")
+    created_at: datetime = Field(
+        description="The date and time the job was created")
+    title: str | None = Field(description="The title of the scraped page")
+    content: str | None = Field(description="The content of the scraped page")
+    error_message: str | None = Field(
+        description="The error message if the scraping failed")
+
+
+class ScrapingResult(BaseModel):
+    total_requests: int = Field(
+        description="The total number of requests made")
+    successful_requests: int = Field(
+        description="The number of successful requests")
+    failed_requests: int = Field(
+        description="The number of failed requests")
+    failed_urls: list[str] = Field(
+        description="The URLs that failed to be scraped")
+    successful_urls: list[str] = Field(
+        description="The URLs that were successfully scraped")
 
 
 @dataclass
@@ -35,15 +83,3 @@ class ChatModelSettings:
     temperature: float = 0.7
     top_p: float = 0.95
     max_tokens: int | None = None
-
-
-class ChatModelProvider:
-    def can_handle(self, chat_model_settings: ChatModelSettings) -> bool:
-        raise NotImplementedError
-
-    def get_chat_model(
-        self,
-        chat_model_settings: ChatModelSettings,
-        verbose: bool = False,
-    ) -> BaseChatModel:
-        raise NotImplementedError
