+++ app/tests/builders/web_scrape_builder.py
@@ -3,7 +3,7 @@
 from datetime import datetime
 from typing import Any
 
-from app.core.web_scrape import ScrapePageResult
+from app.core.domain import ScrapePageResult
 
 
 class MockResponse:
