+++ app/tests/core/test_web_search.py
@@ -2,7 +2,8 @@ from datetime import datetime
 
 import pytest
 
-from app.core.web_search import SearchEngine, SearchProvider
+from app.core.domain import SearchProvider
+from app.core.web_search import SearchEngine
 from app.tests.builders.build import Build
 
 
@@ -25,8 +26,8 @@ def test_search_result_optional_fields():
 
 def test_search_provider_values():
     """Test SearchProvider enum values"""
-    assert SearchProvider.GOOGLE.value == "google"
-    assert SearchProvider.DUCKDUCKGO.value == "duckduckgo"
+    assert SearchProvider.GOOGLE == "google"
+    assert SearchProvider.DUCKDUCKGO == "duckduckgo"
 
 
 @pytest.mark.asyncio
